修改退货单据  可退数量可为小数   ；处理全项目的int float  double类型转换安全问题

+++ app/src/main/java/com/eshop/jinxiaocun/othermodel/bean/ReturnedPurchaseResult.java
@@ -10,17 +10,17 @@ public class ReturnedPurchaseResult {
     private String flow_no;//单据号
     private String item_no;//商品编码
     private String item_name;//商品名称
-    private int sale_qnty;//数量
+    private float sale_qnty;//数量
     private float sale_price;//销售价
     private float sale_money;//销售金额
     private float source_price;//源价
     private String sell_way;//-销售方式: A销售 B退货 D赠送
-    private int re_qty;//可退数量
-    private int already_re_qty; //已退数量
+    private float re_qty;//可退数量
+    private float already_re_qty; //已退数量
     private String item_barcode;//批次号
     private String produce_date;//生产期
     private String valid_date;//有效期
-    private int rp_Qty;//退货数量
+    private float rp_Qty;//退货数量
 
     public String getFlow_no() {
         return flow_no;
@@ -46,11 +46,11 @@ public class ReturnedPurchaseResult {
         this.item_name = item_name;
     }
 
-    public int getSale_qnty() {
+    public float getSale_qnty() {
         return sale_qnty;
     }
 
-    public void setSale_qnty(int sale_qnty) {
+    public void setSale_qnty(float sale_qnty) {
         this.sale_qnty = sale_qnty;
     }
 
@@ -86,19 +86,19 @@ public class ReturnedPurchaseResult {
         this.sell_way = sell_way;
     }
 
-    public int getRe_qty() {
+    public float getRe_qty() {
         return re_qty;
     }
 
-    public void setRe_qty(int re_qty) {
+    public void setRe_qty(float re_qty) {
         this.re_qty = re_qty;
     }
 
-    public int getAlready_re_qty() {
+    public float getAlready_re_qty() {
         return already_re_qty;
     }
 
-    public void setAlready_re_qty(int already_re_qty) {
+    public void setAlready_re_qty(float already_re_qty) {
         this.already_re_qty = already_re_qty;
     }
 
@@ -126,11 +126,11 @@ public class ReturnedPurchaseResult {
         this.valid_date = valid_date;
     }
 
-    public int getRp_Qty() {
+    public float getRp_Qty() {
         return rp_Qty;
     }
 
-    public void setRp_Qty(int rp_Qty) {
+    public void setRp_Qty(float rp_Qty) {
         this.rp_Qty = rp_Qty;
     }
 }
